project2_image_deploy/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
scikit-learn
python-multipart
